--- a/lib/SqlString.js
+++ b/lib/SqlString.js
@@ -17,13 +17,7 @@ var CHARS_ESCAPE_MAP    = {
 
 SqlString.escapeId = function escapeId(val, forbidQualified) {
   if (Array.isArray(val)) {
-    var sql = '';
-
-    for (var i = 0; i < val.length; i++) {
-      sql += (i === 0 ? '' : ', ') + SqlString.escapeId(val[i], forbidQualified);
-    }
-
-    return sql;
+    return val.map(item => SqlString.escapeId(item, forbidQualified)).join(", ");
   } else if (forbidQualified) {
     return '`' + String(val).replace(ID_GLOBAL_REGEXP, '``') + '`';
   } else {