--- a/lib/SqlString.js
+++ b/lib/SqlString.js
@@ -25,6 +25,34 @@ SqlString.escapeId = function escapeId(val, forbidQualified) {
   }
 };
 
+SqlString.escapeIdFunctional = function escapeId(val, forbidQualified) {
+  if (Array.isArray(val)) {
+    return val.map(function (item) {
+      return SqlString.escapeIdFunctional(item, forbidQualified);
+    }).join(', ');
+  } else if (forbidQualified) {
+    return '`' + String(val).replace(ID_GLOBAL_REGEXP, '``') + '`';
+  } else {
+    return '`' + String(val).replace(ID_GLOBAL_REGEXP, '``').replace(QUAL_GLOBAL_REGEXP, '`.`') + '`';
+  }
+};
+
+SqlString.escapeIdIterative = function escapeId(val, forbidQualified) {
+  if (Array.isArray(val)) {
+    var sql = new Array(val.length);
+
+    for (var i = 0; i < val.length; i++) {
+      sql[SqlString.escapeIdIterative(val[i], forbidQualified)];
+    }
+
+    return sql.join(', ');
+  } else if (forbidQualified) {
+    return '`' + String(val).replace(ID_GLOBAL_REGEXP, '``') + '`';
+  } else {
+    return '`' + String(val).replace(ID_GLOBAL_REGEXP, '``').replace(QUAL_GLOBAL_REGEXP, '`.`') + '`';
+  }
+};
+
 SqlString.escape = function escape(val, stringifyObjects, timeZone) {
   if (val === undefined || val === null) {
     return 'NULL';