I implemented functional and iterative version of `SqlString.escapeId`

Here are the benchmarks I got with `npm run bench` on my computer:

```benchmarks
  escapeId

  12 tests completed.

  "col"                            x 1,881,330 ops/sec ±1.37% (189 runs sampled)
  "tbl.col"                        x 1,629,198 ops/sec ±1.38% (190 runs sampled)
  ["col1", "col2"]                 x   707,386 ops/sec ±1.54% (187 runs sampled)
  Array of 20 strings              x    86,097 ops/sec ±1.49% (190 runs sampled)
  "col" <Functional>               x 1,888,515 ops/sec ±1.88% (191 runs sampled)
  "tbl.col" <Functional>           x 1,645,613 ops/sec ±1.28% (191 runs sampled)
  ["col1", "col2"] <Functional>    x   714,820 ops/sec ±1.41% (189 runs sampled)
  Array of 20 strings <Functional> x    86,167 ops/sec ±1.40% (190 runs sampled)
  "col" <Iterative>                x 1,932,067 ops/sec ±1.26% (189 runs sampled)
  "tbl.col" <Iterative>            x 1,660,778 ops/sec ±1.51% (190 runs sampled)
  ["col1", "col2"] <Iterative>     x   632,021 ops/sec ±1.87% (188 runs sampled)
  Array of 20 strings <Iterative>  x    69,959 ops/sec ±1.39% (190 runs sampled)
```

Let me know if you want to merge iterative implementation, and I'll cleanup the code.