@dougwilson 2020-03-08
Can you elaborate more? Based on the posted benchmarks, it would seem that the iterative implementation quickly gets worse for a small margin of being fast in the first two cases?