@michaelkruglos 2020-03-10
I must have misread the numbers. Sorry.
This code and the benchmark results are very counterintuitive.